add script to plot single subject morphometric metrics together with normative values
